+++ .zshrc
@@ -1,0 +1,1 @@
+export PATH="/usr/local/opt/llvm/bin:$PATH"
