+++ Makefile
@@ -7,7 +7,8 @@ LEAK = -fsanitize=leak
 SRCS = pipe/pipe.c pipe/is_x.c pipe/utils_a.c pipe/main.c builtins/cd.c\
 		builtins/echo.c builtins/export.c builtins/pwd.c builtins/unset.c \
 		builtins/env.c read_command.c get_exec_path.c utils.c argument.c\
-		main.c test.c make_strb.c syntax_check.c builtins/exit.c
+		main.c quotes.c make_strb.c syntax_check.c builtins/exit.c pipe/utils_b.c\
+		utils_c.c quote.c
 OBJS = $(SRCS:.c=.o)
 
 all: $(NAME)
