+++ builtins/pwd.c
@@ -1,7 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pwd.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: syudai <syudai@student.42tokyo.jp>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/06 17:52:06 by syudai            #+#    #+#             */
+/*   Updated: 2021/02/06 17:52:21 by syudai           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 #define MAX_FILENAME 1024
 
-int ft_pwd(void)
+int		ft_pwd(void)
 {
 	char s[MAX_FILENAME];
