+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: syudai <syudai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 15:25:45 by syudai            #+#    #+#             */
-/*   Updated: 2021/02/04 23:07:46 by syudai           ###   ########.fr       */
+/*   Updated: 2021/02/06 17:02:22 by syudai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,6 @@
 # include <stdio.h>
 # include <fcntl.h>
 # include <dirent.h>
-//# include <sys/wait.h>
 
 typedef struct	s_list
 {
